+++ CHANGELOG.md
@@ -25,6 +25,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Add new basic tests for the realty object model and mapper (#179)
 
 ### Changed
+- Stop using markTableAsDirty in the new functional tests (#232)
 - Move the RealURL configuration to Classes/ and namespace it (#228, #229)
 - Use fewer DB requests when accessing the realty model (#202, #203)
 - Rename "fixture" to "subject" in the tests (#199)